Running Python Script and various expressions in an interactive interpreter
a.	Create a python program to enter two numbers and then performs and displays the results of the following operations: addition, subtraction, multiplication, and division.
b.	Create a python program to enter two numbers and then performs and displays the results of the following relational expression : >, <, ==, !=, >=, <=

c.	Create a python program to enter three numbers and then performs and displays the results of the following Logical operations: and, or, not.